limit wait time for -1 socket timeouts and increase polling interval
This patch increases the polling frequency to SOCKET_WAIT_MILLI_SECOND even when
passing -1 as the timeout (which is done in all default cases).
As a side effect, waitOnStatus with a -1 timeout does not wait infinitely long
anymore but rather waits for a maximum of SOCKET_STATUS_TIMEOUT_MILLI_SECOND which
is what the name of the constant suggests.